backend/src/beautiful_photometry/beautiful_photometry/requirements.txt
jupyter
colour-science
matplotlib